+++ CHANGELOG.md
@@ -85,6 +85,7 @@
 - Register Web Facet for `acceleratoraddon` [#583](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/583) 
 - Do not register `hmc` sub-modules if there is no `hmc` extension [#582](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/582) 
 - DB Datasource cannot be configured due backslash [#584](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/584) 
+- Some classes are not visible from `storefrontcommons` if searchable class is in read-only storefrontcommons  [#593](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/593)
 
 ### Other
 - Adjusted inline documentation for Type System [#539](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/539)